feat: order queue, caching...

+++ migrations/20230925185748_create_addon_table.ts
@@ -3,7 +3,7 @@ import { Knex } from 'knex';
 export async function up(knex: Knex): Promise<void> {
   return knex.schema.createTable('addons', (table) => {
     table.increments('id').primary();
-    table.decimal('amount');
+    table.float('amount');
     table.integer('meal_id');
   });
 }
